tempfile management, without copy-pasta
