+++ src/xg.hpp
@@ -6,13 +6,11 @@
 #include <queue>
 #include <omp.h>
 #include <unordered_map>
-#include <unordered_set>
 #include <string>
 #include <functional>
 #include <utility>
 #include <tuple>
 #include <sys/types.h>
-#include <dirent.h>
 
 #include <sdsl/bit_vectors.hpp>
 #include <sdsl/enc_vector.hpp>
@@ -31,7 +29,6 @@
 #include <handlegraph/serializable_handle_graph.hpp>
 
 #include <mmmultimap.hpp>
-
 
 namespace xg {
 
@@ -580,32 +577,6 @@ public:
     size_t handle_start(size_t offset) const;
     size_t step_rank_at_position(size_t pos) const;
 };
-
-/**
- * Temporary files. Create with create() and remove with remove(). All
- * temporary files will be deleted when the program exits normally or with
- * std::exit(). The files will be created in a directory determined from
- * environment variables, though this can be overridden with set_dir().
- * The interface is thread-safe.
- */
-namespace temp_file {
-
-/// Create a temporary file starting with the given base name
-std::string create(const std::string& base);
-
-/// Create a temporary file
-std::string create();
-
-/// Remove a temporary file
-void remove(const std::string& filename);
-
-/// Set a temp dir, overriding system defaults and environment variables.
-void set_dir(const std::string& new_temp_dir);
-
-/// Get the current temp dir
-std::string get_dir();
-
-} // namespace temp_file
 
 /// Uses OMP to get the count of threads
 int get_thread_count(void);
